Module scrypto_test::prelude::btree_set

Macros

btreeset

Structs

BTreeSet

An ordered set based on a B-Tree.

Difference

A lazy iterator producing elements in the difference of BTreeSets.

Intersection

A lazy iterator producing elements in the intersection of BTreeSets.

IntoIter

An owning iterator over the items of a BTreeSet.

Iter

An iterator over the items of a BTreeSet.

Range

An iterator over a sub-range of items in a BTreeSet.

SymmetricDifference

A lazy iterator producing elements in the symmetric difference of BTreeSets.

Union

A lazy iterator producing elements in the union of BTreeSets.

ExtractIfExperimental

An iterator produced by calling extract_if on BTreeSet.