Module scrypto_test::prelude::costing

Structs

CostingModule

FeeReserveFinalizationSummary

FeeTable

Fee table specifies how each costing entry should be costed.

NATIVE_FUNCTION_BASE_COSTS

NATIVE_FUNCTION_BASE_COSTS_SIZE_DEPENDENT

SystemLoanFeeReserve

Enums

CostingError

ExecutionCostingEntry

FeeReserveError

FinalizationCostingEntry

OnApplyCost

RoyaltyRecipient

StorageType

Traits

ExecutionFeeReserve

FeeReserve

FinalizingFeeReserve

PreExecutionFeeReserve

This is only allowed before a transaction properly begins. After any other methods are called, this cannot be called again.

Functions

apply_royalty_cost