Enum scrypto_test::prelude::RuntimeError

pub enum RuntimeError {
    KernelError(KernelError),
    SystemError(SystemError),
    SystemModuleError(SystemModuleError),
    SystemUpstreamError(SystemUpstreamError),
    VmError(VmError),
    ApplicationError(ApplicationError),
}

Represents an error when executing a transaction.

Variants

KernelError(KernelError)

An error occurred within the kernel.

SystemError(SystemError)

An error occurred within the system, notably the ClientApi implementation.

SystemModuleError(SystemModuleError)

An error occurred within a specific system module, like auth, costing and royalty. TODO: merge into SystemError?

SystemUpstreamError(SystemUpstreamError)

An error issued by the system when invoking upstream (such as blueprints, node modules). TODO: merge into SystemError?

VmError(VmError)

An error occurred in the vm layer

ApplicationError(ApplicationError)

An error occurred within application logic, like the RE models.