Struct scrypto_test::prelude::radix_engine_common::math::precise_decimal::PreciseDecimal

pub struct PreciseDecimal(pub I256);

PreciseDecimal represents a 256 bit representation of a fixed-scale decimal number.

The finite set of values are of the form m / 10^36, where m is an integer such that -2^(256 - 1) <= m < 2^(256 - 1).

Fractional part: ~120 bits / 36 digits Integer part : 136 bits / 41 digits Max : 57896044618658097711785492504343953926634.992332820282019728792003956564819967 Min : -57896044618658097711785492504343953926634.992332820282019728792003956564819968

Unless otherwise specified, all operations will panic if underflow/overflow.

Tuple Fields
0: I256
Implementations

impl PreciseDecimal

pub const MIN: PreciseDecimal = _

The min value of PreciseDecimal.

pub const MAX: PreciseDecimal = _

The max value of PreciseDecimal.

pub const BITS: usize = 256usize

The bit length of number storing PreciseDecimal.

pub const SCALE: u32 = 36u32

The fixed scale used by PreciseDecimal.

pub const ZERO: PreciseDecimal = _

pub const ONE_HUNDREDTH: PreciseDecimal = _

pub const ONE_TENTH: PreciseDecimal = _

pub const ONE: PreciseDecimal = _

pub const TEN: PreciseDecimal = _

pub const ONE_HUNDRED: PreciseDecimal = _

pub fn zero() -> PreciseDecimal

Returns PreciseDecimal of 0.

pub fn one() -> PreciseDecimal

Returns PreciseDecimal of 1.

pub fn is_zero(&self) -> bool

Whether this decimal is zero.

pub fn is_positive(&self) -> bool

Whether this decimal is positive.

pub fn is_negative(&self) -> bool

Whether this decimal is negative.

pub fn checked_abs(&self) -> Option<PreciseDecimal>

Returns the absolute value.

pub fn checked_floor(&self) -> Option<PreciseDecimal>

Returns the largest integer that is equal to or less than this number.

pub fn checked_ceiling(&self) -> Option<PreciseDecimal>

Returns the smallest integer that is equal to or greater than this number.

pub fn checked_round<T>( &self, decimal_places: T, mode: RoundingMode ) -> Option<PreciseDecimal>
where T: Into<i32>,

Rounds this number to the specified decimal places.

Panics

Panic if the number of decimal places is not within [0..SCALE]

pub fn checked_powi(&self, exp: i64) -> Option<PreciseDecimal>

Calculates power using exponentiation by squaring.

pub fn checked_sqrt(&self) -> Option<PreciseDecimal>

Square root of a PreciseDecimal

pub fn checked_cbrt(&self) -> Option<PreciseDecimal>

Cubic root of a PreciseDecimal

pub fn checked_nth_root(&self, n: u32) -> Option<PreciseDecimal>

Nth root of a PreciseDecimal

impl PreciseDecimal

pub fn to_vec(&self) -> Vec<u8>