Module scrypto_test::prelude::radix_engine_common::prelude

Each module should have its own prelude, which:

Adds preludes of upstream crates

Exports types with specific-enough names which mean they can safely be used downstream.

The idea is that we can just include the current crate’s prelude and avoid messing around with tons of includes. This makes refactors easier, and makes integration into the node less painful.

Modules

basic

SBOR basic, no custom types

btree_map

btree_set

categorize

SBOR Categorize trait

codec

SBOR codec for core Rust types.

constants

SBOR constants

decode

SBOR decode trait.

decoder

SBOR decoding.

encode

SBOR encode trait.

encoded_wrappers

SBOR payload wrappers. These are new types around an encoded payload or sub-payload, with helper methods / traits implemented. They can be used as a more efficient wrapper a ScryptoValue if the content of that value is not needed.

encoder

SBOR encoding.

fmt

Utilities for formatting and printing Strings.

hash_map

hash_set

index_map

The methods and macros provided directly in this index_map module (new, with_capacity) work in both std and no-std modes - unlike the corresponding methods on IndexMap itself.

index_set

The methods and macros provided directly in this index_set module (new, with_capacity) work in both std and no-std modes - unlike the corresponding methods on IndexSet itself.

indexmap

IndexMap is a hash table where the iteration order of the key-value pairs is independent of the hash values of the keys.

non_iter_map

paste

github crates-io docs-rs

path

SBOR paths.

payload_validation

SBOR payload validation.

prelude

Each module should have its own prelude, which:

representations

SBOR textual representations

rust

A facade of Rust types.

schema

SBOR Schema

traversal

SBOR structured payload traversal.

value

SBOR value model and any decoding/encoding.

value_kind

SBOR value kinds - ie the types of value that are supported.

vec

A contiguous growable array type with heap-allocated contents, written Vec<T>.

versioned

Data model versioning helper macro

Macros

btreemap

btreeset

create_well_known_lookup

define_single_versioned

This macro is intended for creating a data model which supports versioning. This is useful for creating an SBOR data model which can be updated in future. In future, enum variants can be added, and automatically mapped to.

define_versioned

This macro is intended for creating a data model which supports versioning. This is useful for creating an SBOR data model which can be updated in future. In future, enum variants can be added, and automatically mapped to.

define_wrapped_hash

format

Creates a String using interpolation of runtime expressions.

hashmap

hashset

i

Creates a safe integer from literals. You must specify the type of the integer you want to create.

indexmap

indexset

look_up_type

manifest_args

numeric_validation_match

return_if_error

return_type_mismatch_error

scrypto_args

terminal_value

terminal_value_from_body

to_manifest_value_and_unwrap

vec

Creates a Vec containing the arguments.

Structs

AddressBech32Decoder

Represents a decoder which understands how to decode Scrypto addresses in Bech32.

AddressBech32Encoder

Represents an encoder which understands how to encode Scrypto addresses in Bech32.

AddressDisplayContext

ArrayData

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

Bls12381G1PrivateKey

Bls12381G1PublicKey

Represents a BLS12-381 G1 public key.

Bls12381G2Signature

Represents a BLS12-381 G2 signature (variant with 96-byte signature and 48-byte public key)

BlueprintTypeIdentifier

BorrowError

An error returned by RefCell::try_borrow.

BorrowMutError

An error returned by RefCell::try_borrow_mut.

Box

A pointer type that uniquely owns a heap allocation of type T.

BytesNonFungibleLocalId

Bytes, of length between 1 and 64.

Cell

A mutable memory location.

ComponentAddress

Address to a global component

ContextDisplayable

ContextSerializable

Decimal

Decimal represents a 192 bit representation of a fixed-scale decimal number.

Ed25519PrivateKey

Ed25519PublicKey

Represents an ED25519 public key.

Ed25519PublicKeyHash

Ed25519Signature

Represents an ED25519 signature.

EnumVariantData

EnumVariantValue

Epoch

A type-safe consensus epoch number.

FixedEnumVariant

FullyScopedTypeId

An identifier for a type under a given node’s schema context in the Radix network.

GlobalAddress

Address to a global entity

GlobalAddressReservation

Hash

Represents a 32-byte hash digest.

HashAccumulator

Represents a 32-byte hash accumulator.

HrpSet

Represents an HRP set (typically corresponds to a network).

I192

Provides safe integer arithmetic.

I256

Provides safe integer arithmetic.

I320

Provides safe integer arithmetic.

I384

Provides safe integer arithmetic.

I448

Provides safe integer arithmetic.

I512

Provides safe integer arithmetic.

I768

Provides safe integer arithmetic.

Instant

Represents a Unix timestamp, capturing the seconds since the unix epoch.

IntegerNonFungibleLocalId

Unsigned integers, up to u64.

InternalAddress

Address to a local entity

LengthValidation

Represents additional validation that should be performed on the size.

LinkedList

A doubly-linked list with owned nodes.

LocatedValidationError

ManifestAddressReservation

ManifestArgs

ManifestBlobRef

ManifestBucket

ManifestCustomTerminalValueRef

ManifestDecimal

ManifestPreciseDecimal

ManifestProof

ManifestValueDisplayContext

Note - this is quite similar to ManifestDecompilationDisplayContext

MapData

NetworkDefinition

Network Definition is intended to be the actual definition of a network

NodeId

The unique identifier of a (stored) node.

NonIterMap

A thin wrapper around a HashMap, which guarantees that a HashMap usage will not result in a non-deterministic execution (simply by disallowing the iteration over its elements).

NovelTypeMetadata

This is the struct used in the Schema

NumericValidation

Represents additional validation that should be performed on the numeric value.

OnceCell

A cell which can be written to only once.

Own

PackageAddress

Address to a global package

PartitionNumber

The unique identifier of a node module.

PartitionOffset

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

PreciseDecimal

PreciseDecimal represents a 256 bit representation of a fixed-scale decimal number.

RUIDNonFungibleLocalId

RUID, v4, variant 1, big endian. See https://www.rfc-editor.org/rfc/rfc4122

RawPayload

A wrapper for a full encoded SBOR payload, including the prefix byte.

RawValue

A wrapper for a reference to a valid partial SBOR payload representing a single value.

Rc

A single-threaded reference-counting pointer. ‘Rc’ stands for ‘Reference Counted’.

Ref

Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell

A mutable memory location with dynamically checked borrow rules

RefMut

A wrapper type for a mutably borrowed value from a RefCell<T>.

Reference

ResourceAddress

Address to a global resource

Round

A type-safe consensus round number within a single epoch.

SborFixedEnumVariant

SborPath

A series of indexes which describes some value in the sbor tree

SborPathBuf

SborRawPayload

A wrapper for a full encoded SBOR payload, including the prefix byte.

SborRawValue

A wrapper for a reference to a valid partial SBOR payload representing a single value.

SchemaContext

SchemaHash

Represents a particular schema under a package

SchemaV1

An array of custom type kinds, and associated extra information which can attach to the type kinds

ScopedTypeId

An identifier for a type in the context of a schema.

ScryptoCustomExtension

ScryptoCustomSchema

ScryptoCustomTerminalValueRef

ScryptoValueDisplayContext

Secp256k1PrivateKey

Secp256k1PublicKey

Represents an ECDSA Secp256k1 public key.

Secp256k1PublicKeyHash

Secp256k1Signature

Represents an ECDSA Secp256k1 signature.

String

A UTF-8–encoded, growable string.

StringNonFungibleLocalId

A string matching [_0-9a-zA-Z]{1,64}.

TupleData

TypeAggregator

TypeData

Combines all data about a Type:

TypeMetadata

This enables the type to be represented as eg JSON Also used to facilitate type reconstruction

U192

Provides safe integer arithmetic.

U256

Provides safe integer arithmetic.

U320

Provides safe integer arithmetic.

U384

Provides safe integer arithmetic.

U448

Provides safe integer arithmetic.

U512

Provides safe integer arithmetic.

U768

Provides safe integer arithmetic.

UnsafeCell

The core primitive for interior mutability in Rust.

UtcDateTime

A UtcDateTime represents a Unix timestamp on the UTC Calendar.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

VecDecoder

A Decoder abstracts the logic for decoding basic types.

VecDeque

A double-ended queue implemented with a growable ring buffer.

LazyCellExperimental

A value which is initialized on the first access.

SyncUnsafeCellExperimental

UnsafeCell, but Sync.

VecEncoder

An Encoder abstracts the logic for writing core types into a byte buffer.

WellKnownTypeId

Enums

AddressBech32DecodeError

AddressBech32EncodeError

ChildNames

ContentValidationError

Cow

A clone-on-write smart pointer.

DateTimeError

DecodeError

Represents an error occurred during decoding.

EncodeError

Represents an error occurred during encoding.

EntityType

An enum which represents the different addressable entities.

GenericSubstitution

A reference to the type to substitute with for the case of generics.

LocalTypeId

The TypeId which is local to a given Schema. This is the SchemaTypeLink used in a linearized Schema to link TypeKinds.

ManifestAddress

Any address supported by manifest, both global and local.

ManifestCustomExtension

ManifestCustomTraversal

ManifestCustomValue

ManifestCustomValueKind

ManifestExpression

ManifestNonFungibleLocalId

ManifestNonFungibleLocalIdValidationError

ManifestToRustValueError

NoCustomExtension

NoCustomSchema

NoCustomTerminalValueRef

NoCustomTraversal

NoCustomTypeKind

NoCustomTypeValidation

NoCustomValue

NoCustomValueKind

NonFungibleIdType

Represents type of non-fungible id

NonFungibleLocalId

Represents the local id of a non-fungible.

Option

The Option type. See the module level documentation for more.

OwnValidation

ParseBlsPublicKeyError

Represents an error when retrieving BLS public key from hex or when aggregating.

ParseBlsSignatureError

Represents an error when retrieving BLS signature from hex or when aggregating.

ParseComponentAddressError

ParseDecimalError

Represents an error when parsing Decimal from another type.

ParseEd25519PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseEd25519SignatureError

ParseGlobalAddressError

ParseHashError

Represents an error when parsing hash.

ParseI192Error

ParseI256Error

ParseI320Error

ParseI384Error

ParseI448Error

ParseI512Error

ParseI768Error

ParseInternalAddressError

ParseManifestAddressReservationError

Represents an error when parsing ManifestAddressReservation.

ParseManifestBlobRefError

Represents an error when parsing ManifestBlobRef.

ParseManifestBucketError

Represents an error when parsing ManifestBucket.

ParseManifestDecimalError

Represents an error when parsing ManifestDecimal.

ParseManifestExpressionError

Represents an error when parsing ManifestExpression.

ParseManifestPreciseDecimalError

Represents an error when parsing ManifestPreciseDecimal.

ParseManifestProofError

Represents an error when parsing ManifestProof.

ParseNetworkError

ParseNonFungibleLocalIdError

Represents an error when decoding non-fungible id.

ParseOwnError

ParsePackageAddressError

ParsePreciseDecimalError

Represents an error when parsing PreciseDecimal from another type.

ParseReferenceError

ParseResourceAddressError

ParseSecp256k1PublicKeyError

Represents an error when parsing ED25519 public key from hex.

ParseSecp256k1SignatureError

ParseU192Error

ParseU256Error

ParseU320Error

ParseU384Error

ParseU448Error

ParseU512Error

ParseU768Error

ParseUtcDateTimeError

PayloadValidationError

PublicKey

Represents any natively supported public key.

PublicKeyHash

The hash of a given public key.

ReferenceValidation

Result

Result is a type that represents either success (Ok) or failure (Err).

RoundingMode

Defines the rounding strategy.

RoyaltyAmount

RustToManifestValueError

RustTypeId

This is a compile-time identifier for a given type, used by the type aggregator to uniquely identify a type.

SborValue

Represents any value conforming to the SBOR value model.

SchemaValidationError

ScryptoCustomTraversal

ScryptoCustomTypeKind

A schema for the values that a codec can decode / views as valid

ScryptoCustomTypeValidation

ScryptoCustomValue

ScryptoCustomValueKind

SubstateKey

The unique identifier of a substate within a node module.

TimeComparisonOperator

TypeKind

A schema for the values that a codec can decode / views as valid

TypeValidation

Additional validation to apply to a payload of the given type, beyond validation from the TypeKind’s type structure.

UpdateResult

ValidationError

Value

Represents any value conforming to the SBOR value model.

ValueKind

VersionedSchema

Constants

ACCESS_CONTROLLER_PACKAGE

The native package for access controllers.

ACCESS_RULE_NODE_LIST_TYPE

ACCESS_RULE_NODE_TYPE

ACCESS_RULE_TYPE

ACCOUNT_OWNER_BADGE

The non-fungible badge resource which is used for account ownership, if accounts have been set up with simple account creation, or have been securified.

ACCOUNT_PACKAGE

The native package for accounts.

ARCHIVE_STORAGE_PRICE_IN_XRD

The price for adding a single byte to archive storage, in XRD. 1 MB = 6 USD This is primarily for transaction payload, events and logs.

ATTACHED_MODULE_ID_TYPE

BASIC_SBOR_V1_MAX_DEPTH

BASIC_SBOR_V1_PAYLOAD_PREFIX

BLS12381_CIPHERSITE_V1

BLS12-381 ciphersuite v1 It has following parameters

BLUEPRINT_PAYLOAD_MAX_DEPTH

Depth limit for various types of blueprint payload

BOOT_LOADER_RESERVED_NODE_ID_FIRST_BYTE

COMPONENT_ADDRESS_TYPE

CONSENSUS_MANAGER

The consensus manager native component - in charge of validators, consensus and epochs.

CONSENSUS_MANAGER_PACKAGE

The native package for the consensus manager.

CUSTOM_VALUE_KIND_START

A custom type is an application defined type with special semantics.

DECIMAL_SIZE

DECIMAL_TYPE

DEFAULT_TIP_PERCENTAGE

The default tip percentage to use when building a transaction.

ED25519_PUBLIC_KEY_HASH_TYPE

ED25519_PUBLIC_KEY_TYPE

ED25519_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of EdDSA Ed25519 transacton signatures in the transaction processor.

EXECUTION_COST_UNIT_LIMIT

The execution cost unit limit.

EXECUTION_COST_UNIT_LOAN

The execution cost units loaned from system

EXECUTION_COST_UNIT_PRICE_IN_XRD

The price of execution cost unit, in XRD.

FAUCET

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_BLUEPRINT

The name of the faucet blueprint under the FAUCET_PACKAGE.

FAUCET_COMPONENT

The faucet native component - use this on testnets for getting XRD and locking fee.

FAUCET_PACKAGE

The scrypto package for the faucet

FINALIZATION_COST_UNIT_LIMIT

The finalization cost unit limit.

FINALIZATION_COST_UNIT_PRICE_IN_XRD

The price of finalization cost unit, in XRD.

GENESIS_HELPER

The genesis helper scrypto component - used for sorting out genesis.

GENESIS_HELPER_BLUEPRINT

The name of the genesis helper blueprint under the GENESIS_HELPER_PACKAGE.

GENESIS_HELPER_PACKAGE

The scrypto package for the genesis helper.

GLOBAL_ADDRESS_TYPE

GLOBAL_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the global ancestor of the actor which made the latest global call.

IDENTITY_OWNER_BADGE

The non-fungible badge resource which is used for identity ownership, if identities have been set up with simple account creation, or have been securified.

IDENTITY_PACKAGE

The native package for identities.

INSTANT_TYPE

INTERNAL_ADDRESS_TYPE

KEY_VALUE_STORE_PAYLOAD_MAX_DEPTH

Depth limit for the key and value of an entry in KeyValueStore

MANIFEST_ADDRESS_DISCRIMINATOR_NAMED

MANIFEST_ADDRESS_DISCRIMINATOR_STATIC

MANIFEST_NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

MANIFEST_SBOR_V1_MAX_DEPTH

MANIFEST_SBOR_V1_PAYLOAD_PREFIX

0x4d = M in ASCII for Manifest - (77 in decimal)

MANIFEST_VALUE_KIND_ADDRESS

MANIFEST_VALUE_KIND_ADDRESS_RESERVATION

MANIFEST_VALUE_KIND_BLOB

MANIFEST_VALUE_KIND_BUCKET

MANIFEST_VALUE_KIND_DECIMAL

MANIFEST_VALUE_KIND_EXPRESSION

MANIFEST_VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

MANIFEST_VALUE_KIND_PRECISE_DECIMAL

MANIFEST_VALUE_KIND_PROOF

MAX_ACCESS_RULE_DEPTH

The max depth of an access rule, to protect unbounded native stack usage

MAX_ACCESS_RULE_NODES

The max number of access rule nodes in an access rule

MAX_BLUEPRINT_NAME_LEN

The max length of a blueprint identifier

MAX_CALL_DEPTH

The max call depth, used by transaction executor.

MAX_EPOCH_RANGE

The max epoch range Should be ~ 1 month. The below is ~30 days given 5 minute epochs.

MAX_EVENT_NAME_LEN

The max length of an event name

MAX_EVENT_SIZE

The max event size

MAX_EXECUTION_TRACE_DEPTH

MAX_FEATURE_NAME_LEN

The max length of a feature name

MAX_FUNCTION_NAME_LEN

The max length of a function name

MAX_HEAP_SUBSTATE_TOTAL_BYTES

The max total heap substate size.

MAX_INITIAL_TABLE_SIZE

The maximum initial table size

MAX_INVOKE_PAYLOAD_SIZE

The maximum invoke payload size.

MAX_LOG_SIZE

The max log size

MAX_MEMORY_SIZE_IN_PAGES

The maximum memory size (per call frame): 64 * 64KiB = 4MiB

MAX_METADATA_KEY_STRING_LEN

The max SBOR size of metadata key

MAX_METADATA_VALUE_SBOR_LEN

The max SBOR size of metadata value

MAX_NUMBER_OF_BLOBS

MAX_NUMBER_OF_BLUEPRINT_FIELDS

The max number of fields a blueprint can have

MAX_NUMBER_OF_BR_TABLE_TARGETS

The max number of labels of a table jump, excluding the default

MAX_NUMBER_OF_BUFFERS

The maximum number of “live” buffers maintained by Scrypto runtime.

MAX_NUMBER_OF_EVENTS

The max number of events

MAX_NUMBER_OF_FIELDS

MAX_NUMBER_OF_FUNCTIONS

The max number of functions

MAX_NUMBER_OF_FUNCTION_LOCALS

The max number of function local variables

MAX_NUMBER_OF_FUNCTION_PARAMS

The max number of function parameters

MAX_NUMBER_OF_GLOBALS

The max number of global variables

MAX_NUMBER_OF_INTENT_SIGNATURES

MAX_NUMBER_OF_LOGS

The max number of logs

MAX_ORIGIN_LENGTH

The max length of an Origin in metadata

MAX_PANIC_MESSAGE_SIZE

The max panic message size

MAX_PER_FUNCTION_ROYALTY_IN_XRD

The maximum that a package or component owner is allowed to set their method royalty to. 10 USD

MAX_ROLES

The max number of roles in a Role Specification

MAX_ROLE_NAME_LEN

The max length of a role name

MAX_SUBSTATE_KEY_SIZE

The maximum substate key read and write size

MAX_SUBSTATE_VALUE_SIZE

The maximum substate read and write size.

MAX_TIP_PERCENTAGE

The maximum value of tip percentage

MAX_TRACK_SUBSTATE_TOTAL_BYTES

The max total track substate size.

MAX_TRANSACTION_SIZE

The max transaction size

MAX_TYPE_NAME_LEN

The max length of a registered type name

MAX_URL_LENGTH

The max length of a URL in metadata

METADATA_MODULE_PACKAGE

The native package for the metadata module.

MIN_TIP_PERCENTAGE

The minimum value of tip percentage

MODULE_ID_TYPE

NETWORK_FEES_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of network fees (execution, finalization and storage)

NETWORK_FEES_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of network fees (execution, finalization and storage)

NON_FUNGIBLE_GLOBAL_ID_TYPE

NON_FUNGIBLE_LOCAL_ID_MAX_LENGTH

NON_FUNGIBLE_LOCAL_ID_TYPE

OPTION_VARIANT_NONE

OPTION_VARIANT_SOME

ORIGIN_TYPE

OWNER_ROLE_TYPE

OWN_BUCKET_TYPE

OWN_FUNGIBLE_BUCKET_TYPE

OWN_FUNGIBLE_PROOF_TYPE

OWN_FUNGIBLE_VAULT_TYPE

OWN_GLOBAL_ADDRESS_RESERVATION_TYPE

OWN_KEY_VALUE_STORE_TYPE

OWN_NON_FUNGIBLE_BUCKET_TYPE

OWN_NON_FUNGIBLE_PROOF_TYPE

OWN_NON_FUNGIBLE_VAULT_TYPE

OWN_PROOF_TYPE

OWN_TYPE

OWN_VAULT_TYPE

PACKAGE_ADDRESS_TYPE

PACKAGE_OF_DIRECT_CALLER_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs which represent the package of the immediate caller - ie the actor which made the latest (global or internal) call.

PACKAGE_OWNER_BADGE

The non-fungible badge resource which is used for package ownership when creating packages with the simple package creation set-up.

PACKAGE_PACKAGE

The native package for package deployment.

POOL_PACKAGE

The native package for pools.

PRECISE_DECIMAL_SIZE

PRECISE_DECIMAL_TYPE

PREVIEW_CREDIT_IN_XRD

The free credit amount used by preview. 1000000 XRD = $60000, which should be enough to cover all sensible preview requests.

PROOF_RULE_TYPE

PUBLIC_KEY_HASH_TYPE

PUBLIC_KEY_TYPE

REFERENCE_TYPE

RESOURCE_ADDRESS_TYPE

RESOURCE_OR_NON_FUNGIBLE_LIST_TYPE

RESOURCE_OR_NON_FUNGIBLE_TYPE

RESOURCE_PACKAGE

The native package for resource managers, proofs, buckets, vaults etc.

RESULT_VARIANT_ERR

RESULT_VARIANT_OK

ROLA_HASHABLE_PAYLOAD_PREFIX

The ROLA hash which is signed is created as hash(ROLA_HASHABLE_PAYLOAD_PREFIX || ..)

ROLE_ASSIGNMENT_MODULE_PACKAGE

The native package for the role assignment module.

ROLE_KEY_TYPE

ROYALTY_AMOUNT_TYPE

ROYALTY_MODULE_PACKAGE

The native package for the royalty module.

SCRYPTO_SBOR_V1_MAX_DEPTH

SCRYPTO_SBOR_V1_PAYLOAD_PREFIX

0x5c for [5c]rypto - (92 in decimal)

SECP256K1_PUBLIC_KEY_HASH_TYPE

SECP256K1_PUBLIC_KEY_TYPE

SECP256K1_SIGNATURE_VIRTUAL_BADGE

The non-fungible badge resource which is used for virtual proofs of ECDSA Secp256k1 transacton signatures in the transaction processor.

STATE_STORAGE_PRICE_IN_XRD

The price for adding a single byte to state storage, in XRD. 1 MB = 6 USD

SYSTEM_TRANSACTION_BADGE

The non-fungible badge resource which is used for virtual proofs representing the fact that the current transaction is a system transaction.

TEST_UTILS_PACKAGE

The native package for test utils.

TIPS_PROPOSER_SHARE_PERCENTAGE

The proposer’s share of tips

TIPS_VALIDATOR_SET_SHARE_PERCENTAGE

The validator set’s share of tips

TRANSACTION_HASHABLE_PAYLOAD_PREFIX

The Transaction hash which is signed is created as: hash(TRANSACTION_HASHABLE_PAYLOAD_PREFIX || version prefix according to type of transaction payload || ..)

TRANSACTION_PROCESSOR_PACKAGE

The native package for the transaction processor.

TRANSACTION_TRACKER

The intent hash store component

TRANSACTION_TRACKER_BLUEPRINT

TRANSACTION_TRACKER_PACKAGE

TRANSACTION_VERSION_V1

TRANSIENT_SUBSTATE_DEFAULT_VALUE_MAX_DEPTH

Depth limit for the default value of a transient substate

URL_TYPE

USD_PRICE_IN_XRD

The USD price, in XRD. 1 XRD = 0.06 USD

UTC_DATE_TIME_TYPE

VALIDATOR_OWNER_BADGE

The non-fungible badge resource which is used for validator ownership.

VALUE_KIND_ARRAY

VALUE_KIND_BOOL

VALUE_KIND_DECIMAL

VALUE_KIND_ENUM

VALUE_KIND_I8

VALUE_KIND_I16

VALUE_KIND_I32

VALUE_KIND_I64

VALUE_KIND_I128

VALUE_KIND_MAP

VALUE_KIND_NON_FUNGIBLE_LOCAL_ID

VALUE_KIND_OWN

VALUE_KIND_PRECISE_DECIMAL

VALUE_KIND_REFERENCE

VALUE_KIND_STRING

VALUE_KIND_TUPLE

VALUE_KIND_U8

VALUE_KIND_U16

VALUE_KIND_U32

VALUE_KIND_U64

VALUE_KIND_U128

WASM_ENGINE_CACHE_SIZE

The number of entries in the engine cache

WASM_EXECUTION_COST_UNITS_BUFFER

XRD

XRD is the native token of the Radix ledger. It is a fungible token, measured in attos (10^-18).

Statics

NO_NETWORK

Traits

AsMut

Used to do a cheap mutable-to-mutable reference conversion.

AsRef

Used to do a cheap reference-to-reference conversion.

BasicCategorize

BasicDecode

BasicDescribe

BasicEncode

BasicSbor

BasicSborEnum

BasicSborTuple

BorrowingDecoder

Categorize

The Categorize trait marks a rust type as having a fixed value kind for SBOR encoding/decoding.

Cbrt

CheckedAdd

CheckedDiv

CheckedMul

CheckedNeg

CheckedSub

CheckedTruncate

Clone

A common trait for the ability to explicitly duplicate an object.

CloneIntoLatest

ContextualDisplay

This trait is used where context is required to correctly display a value.

ContextualSerialize

This trait is used where context is required to correctly serialize a value.

Copy

Types whose values can be duplicated simply by copying bits.

CustomExtension

CustomSchema

CustomTypeKind

CustomTypeValidation

CustomValue

Represents a custom SBOR value.

CustomValueKind

Debug

? formatting.

Decode

A data structure that can be decoded from a byte array using SBOR.

Decoder

Default

A trait for giving a type a useful default value.

Describe

The Describe trait allows a type to describe how to interpret and validate a corresponding SBOR payload.

Display

Format trait for an empty format, {}.

DoubleEndedIterator

An iterator able to yield elements from both ends.

Drop

Custom code within the destructor.

Encode

A data structure that can be serialized into a byte array using SBOR.

Encoder

Eq

Trait for comparisons corresponding to equivalence relations.

ExactSizeIterator

An iterator that knows its exact length.

Extend

Extend a collection with the contents of an iterator.

Fn

The version of the call operator that takes an immutable receiver.

FnMut

The version of the call operator that takes a mutable receiver.

FnOnce

The version of the call operator that takes a by-value receiver.

From

Used to do value-to-value conversions while consuming the input value. It is the reciprocal of Into.

FromIterator

Conversion from an Iterator.

FromStr

Parse a value from a string

HasLatestVersion

A marker trait to indicate that the type is versioned. This can be used for type bounds for requiring that types are versioned.

HasPublicKeyHash

HasSchemaHash

Into

A value-to-value conversion that consumes the input value. The opposite of From.

IntoIterator

Conversion into an Iterator.

IsFixedEnumVariant

IsHash

IsNonAutoGeneratedNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of non-auto-generated kind (i.e. String, Integer and Bytes).

IsNonFungibleLocalId

Marks the rust type that represents a non-fungible id, of any kind (i.e. String, Integer, Bytes and RUID).

IsPublicKeyHash

Iterator

A trait for dealing with iterators.

ManifestCategorize

ManifestDecode

ManifestEncode

ManifestSborEnum

ManifestSborTuple

NthRoot

Ord

Trait for types that form a total order.

PartialEq

Trait for comparisons using the equality operator.

PartialOrd

Trait for types that form a partial order.

SborCustomValue

Represents a custom SBOR value.

SborEnum

Marker trait for encoding as an Sbor Enum

SborTuple

Marker trait for encoding as an Sbor Tuple

SchemaTypeLink

Marker trait for a link between TypeKinds:

ScryptoCategorize

ScryptoDecode

ScryptoDescribe

ScryptoEncode

ScryptoSbor

ScryptoSborEnum

ScryptoSborTuple

Send

Types that can be transferred across thread boundaries.

Sized

Types with a constant size known at compile time.

Sqrt

Sync

Types for which it is safe to share references between threads.

ToOwned

A generalization of Clone to borrowed data.

ToString

A trait for converting a value to a String.

TryFrom

Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto

An attempted conversion that consumes self, which may or may not be expensive.

Unpin

Types that do not require any pinning guarantees.

ValidatableCustomExtension

ValueVisitor

VersionedSchemaVersion

Functions

access_rule_node_list_type_data

access_rule_node_type_data

access_rule_type_data

aggregate_verify_bls12381_v1

Performs BLS12-381 G2 aggregated signature verification of multiple messages each signed with different key. Domain specifier tag: BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_

attached_module_id_type_data

basic_decode

Decode an instance of T from a slice.

basic_decode_with_depth_limit

basic_encode

Encode a T into byte array.

basic_encode_with_depth_limit

basic_payload_traverser

Creates a payload traverser from the buffer

bech32_encode_to_fmt

NOTE: The below code is copied with minor alterations from the bech32 crate. These alterations are to avoid using std for allocations, and fit with the sbor no-alloc options.

blake2b_256_hash

combine

Combines a u8 with a u8 slice.

component_address

component_address_type_data

copy_u8_array

Copies a slice to a fixed-sized array.

decimal_type_data

drop

Disposes of a value.

ed25519_public_key_hash_type_data

ed25519_public_key_type_data

fast_aggregate_verify_bls12381_v1

Performs BLS12-381 G2 aggregated signature verification one message signed with multiple keys. Domain specifier tag: BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_

from_manifest_value

generate_full_schema

generate_full_schema_from_single_type

global_address_type_data

hash

Computes the hash digest of a message.

hash_map_new

Creates an empty map with capacity 0 and default Hasher

hash_map_with_capacity

Creates an empty map with given capacity and default Hasher

hash_public_key_bytes

hash_set_new

Creates an empty set with capacity 0 and default Hasher

hash_set_with_capacity

Creates an empty set with given capacity and default Hasher

index_map_new

This is safe for std and no-std use cases (unlike IndexMap::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_map_with_capacity

This is safe for std and no-std use cases (unlike IndexMap::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_new

This is safe for std and no-std use cases (unlike IndexSet::new which disappears when std is not in the toolchain - see this article for deep technical reasons)

index_set_with_capacity

This is safe for std and no-std use cases (unlike IndexSet::with_capacity which disappears when std is not in the toolchain - see this article for deep technical reasons)

instant_type_data

internal_address_type_data

keccak256_hash

local_address

localize_well_known

localize_well_known_type_data

manifest_decode

manifest_decode_with_depth_limit

manifest_encode

manifest_encode_with_depth_limit

module_id_type_data

non_fungible_global_id_type_data

non_fungible_local_id_type_data

origin_type_data

own_bucket_type_data

own_fungible_bucket_type_data

own_fungible_proof_type_data

own_fungible_vault_type_data

own_global_address_reservation_type_data

own_key_value_store_type_data

own_non_fungible_bucket_type_data

own_non_fungible_proof_type_data

own_non_fungible_vault_type_data

own_proof_type_data

own_type_data

own_vault_type_data

owner_role_type_data

package_address

package_address_type_data

precise_decimal_type_data

proof_rule_type_data

public_key_hash_type_data

public_key_type_data

recover_secp256k1

reference_type_data

replace_self_package_address

resolve_local_type_id

resolve_scrypto_well_known_type

resource_address

resource_address_type_data

resource_or_non_fungible_list_type_data

resource_or_non_fungible_type_data

role_key_type_data

royalty_amount_type_data

scrypto_decode

Decodes a data structure from a byte array.

scrypto_decode_with_depth_limit

scrypto_encode

Encodes a data structure into byte array.

scrypto_encode_with_depth_limit

secp256k1_public_key_hash_type_data

secp256k1_public_key_type_data

to_manifest_value

traverse_any

============================================== (DEPRECATED) TRAVERSAL

url_type_data

utc_date_time_type_data

validate_childless_metadata

validate_container

validate_custom_type_validation

validate_enum_metadata

validate_field_names

validate_index

validate_payload_against_schema

validate_schema

validate_schema_enum_variant_name

validate_schema_field_name

validate_schema_type_name

validate_terminal_value

validate_terminal_value_batch

validate_tuple_metadata

validate_type_kind

validate_type_metadata_with_type_kind

verify_bls12381_v1

Performs BLS12-381 G2 signature verification. Domain specifier tag: BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_POP_

verify_ed25519

verify_secp256k1

Type Aliases

BasicDecoder

BasicEncoder

BasicOwnedRawPayload

BasicOwnedRawValue

BasicRawPayload

BasicRawValue

BasicSchema

BasicTraverser

BasicTypeData

BasicTypeKind

BasicValue

BasicValueKind

BasicVersionedSchema

Blake2b256

FieldKey

HashMap

HashSet

IndexMap

IndexSet

ManifestDecoder

ManifestEncoder

ManifestEnumVariantValue

ManifestOwnedRawPayload

ManifestOwnedRawValue

ManifestRawPayload

ManifestRawValue

ManifestTraverser

ManifestValue

ManifestValueKind

MapKey

RawScryptoPayload

RawScryptoValue

Schema

SchemaTypeKind

ScryptoDecoder

ScryptoEncoder

ScryptoOwnedRawPayload

ScryptoOwnedRawValue

ScryptoRawPayload

ScryptoRawValue

ScryptoTraverser

ScryptoTypeData

ScryptoTypeKind

ScryptoValue

ScryptoValueKind

SortedKey

TypeHash

ValidatorIndex

An index of a specific validator within the current validator set. To be exact: a ValidatorIndex equal to k references the k-th element returned by the iterator of the IndexMap<ComponentAddress, Validator> in this epoch’s active validator set (which is expected to be sorted by stake, descending). This uniquely identifies the validator, while being shorter than ComponentAddress (we do care about the constant factor of the space taken by LeaderProposalHistory under prolonged liveness break scenarios).

VersionedScryptoSchema

Derive Macros

BasicCategorize

Derive code that returns the value kind - specifically for Basic SBOR.

BasicDecode

Derive code that decodes this data structure from a byte array - specifically for Basic SBOR.

BasicDescribe

Derive code that describes the type - specifically for Basic SBOR.

BasicEncode

Derive code that encodes this data structure - specifically for Basic SBOR.

BasicSbor

A shortcut for BasicCategorize, BasicEncode, BasicDecode, and BasicDescribe derives.

Categorize

Derive code that returns the value kind.

Clone

Derive macro generating an impl of the trait Clone.

Copy

Derive macro generating an impl of the trait Copy.

Debug

Derive macro generating an impl of the trait Debug.

Decode

Derive code that decodes this data structure from a byte array.

Default

Derive macro generating an impl of the trait Default.

Describe

Derive code that describes this type.

Encode

Derive code that encodes this data structure

Eq

Derive macro generating an impl of the trait Eq.

ManifestCategorize

Derives code for categorizing a struct or enum with Manifest value model.

ManifestDecode

Derives code for decoding a struct or enum with Manifest value model.

ManifestEncode

Derives code for encoding a struct or enum with Manifest value model.

ManifestSbor

A shortcut for ManifestCategorize, ManifestEncode and ManifestDecode derives.

Ord

Derive macro generating an impl of the trait Ord. The behavior of this macro is described in detail here.

PartialEq

Derive macro generating an impl of the trait PartialEq. The behavior of this macro is described in detail here.

PartialOrd

Derive macro generating an impl of the trait PartialOrd. The behavior of this macro is described in detail here.

Sbor

A shortcut for Categorize, Encode, Decode, and Describe derives.

ScryptoCategorize

Derives code for categorizing a struct or enum with Scrypto value model.

ScryptoDecode

Derives code for decoding a struct or enum with Scrypto value model.

ScryptoEncode

Derives code for encoding a struct or enum with Scrypto value model.

ScryptoEvent

Derive code for implementing the required logic to mark a type as being an event.

ScryptoSbor

A shortcut for ScryptoCategorize, ScryptoEncode, ScryptoDecode, and [ScryptoDescribe] derives.