Module scrypto_test::prelude::radix_engine_common::prelude::hash_set

Macros

hashset

Structs

Difference

A lazy iterator producing elements in the difference of HashSets.

Drain

A draining iterator over the items of a HashSet.

Intersection

A lazy iterator producing elements in the intersection of HashSets.

IntoIter

An owning iterator over the items of a HashSet.

Iter

An iterator over the items of a HashSet.

SymmetricDifference

A lazy iterator producing elements in the symmetric difference of HashSets.

Union

A lazy iterator producing elements in the union of HashSets.

ext_HashSet

A hash set implemented as a HashMap where the value is ().

ExtractIfExperimental

A draining, filtering iterator over the items of a HashSet.

Functions

new

Creates an empty set with capacity 0 and default Hasher

with_capacity

Creates an empty set with given capacity and default Hasher

Type Aliases

DefaultHashBuilder

HashSet