Trait scrypto_test::prelude::representations::FormattableCustomExtension

pub trait FormattableCustomExtension: CustomExtension + Copy {
    type CustomDisplayContext<'a>: CustomDisplayContext<'a>;

    // Required method
    fn display_string_content<F, 's, 'de, 'a, 't, 's1, 's2>(
        f: &mut F,
        context: &Self::CustomDisplayContext<'a>,
        value: &<Self::CustomTraversal as CustomTraversal>::CustomTerminalValueRef<'de>
    ) -> Result<(), Error>
       where F: Write;
}

Required Associated Types

type CustomDisplayContext<'a>: CustomDisplayContext<'a>

Required Methods

fn display_string_content<F, 's, 'de, 'a, 't, 's1, 's2>( f: &mut F, context: &Self::CustomDisplayContext<'a>, value: &<Self::CustomTraversal as CustomTraversal>::CustomTerminalValueRef<'de> ) -> Result<(), Error>
where F: Write,

The gives the inner formatted representation of the value. This function should write the value content to the formatter.

The rust-like representaiton is as a newtytpe: CustomValueKind(<value_content>)

The nested string representation is identical: CustomValueKind(<value_content>)

Object Safety

This trait is not object safe.