Struct scrypto_test::prelude::schema::SchemaV1

pub struct SchemaV1<S>
where
    S: CustomSchema,
{
    pub type_kinds: Vec<TypeKind<<S as CustomSchema>::CustomTypeKind<LocalTypeId>, LocalTypeId>>,
    pub type_metadata: Vec<TypeMetadata>,
    pub type_validations: Vec<TypeValidation<<S as CustomSchema>::CustomTypeValidation>>,
}

An array of custom type kinds, and associated extra information which can attach to the type kinds

Fields
type_kinds: Vec<TypeKind<<S as CustomSchema>::CustomTypeKind<LocalTypeId>, LocalTypeId>>type_metadata: Vec<TypeMetadata>type_validations: Vec<TypeValidation<<S as CustomSchema>::CustomTypeValidation>>
Implementations

impl<S> SchemaV1<S>
where S: CustomSchema,

pub fn empty() -> SchemaV1<S>

pub fn resolve_type_kind( &self, type_id: LocalTypeId ) -> Option<&TypeKind<<S as CustomSchema>::CustomTypeKind<LocalTypeId>, LocalTypeId>>

pub fn resolve_type_metadata( &self, type_id: LocalTypeId ) -> Option<&TypeMetadata>

pub fn resolve_matching_tuple_metadata( &self, type_id: LocalTypeId, fields_length: usize ) -> TupleData<'_>

pub fn resolve_matching_enum_metadata<'s>( &self, type_id: LocalTypeId, variant_id: u8, fields_length: usize ) -> EnumVariantData<'_>

pub fn resolve_matching_array_metadata( &self, type_id: LocalTypeId ) -> ArrayData<'_>

pub fn resolve_matching_map_metadata(&self, type_id: LocalTypeId) -> MapData<'_>

pub fn resolve_type_name_from_metadata( &self, type_id: LocalTypeId ) -> Option<&str>

pub fn resolve_type_validation( &self, type_id: LocalTypeId ) -> Option<&TypeValidation<<S as CustomSchema>::CustomTypeValidation>>

pub fn validate(&self) -> Result<(), SchemaValidationError>