Module scrypto_test::prelude::state_updates

Structs

IterationCountedIter

RuntimeSubstate

StateUpdates

A tree-like description of all updates that happened to a stored state, to be included as a part of a transaction receipt. This structure is indexed (i.e. uses IndexMaps where Vecs could be used) for convenience and performance, since both the source (i.e. Track) and the sink (i.e. Database and API) operate on indexed structures too. This structure maintains partial information on the order of operations (please see individual fields for details), since the end users care about it. Please note that this means multiple instances of StateUpdates can represent the same transform of state store (i.e. differing only by order of some operations), and hence it is not 100% “canonical form”.

TrackedNode

TrackedPartition

TrackedSubstate

Enums

BatchPartitionStateUpdate

A description of a batch update affecting an entire Partition.

NodeStateUpdates

A description of all updates that happened to a state of a single Node. Note: currently, we do not support any Node-wide changes (e.g. deleting entire Node); however, we use an enum for potential future development.

PartitionStateUpdates

A description of all updates that happened to a state of a single Partition.

ReadOnly

TrackedSubstateValue

Write

Functions

to_state_updates