Struct scrypto_test::prelude::InstructionWeights

pub struct InstructionWeights {

    pub version: u32,
    pub fallback: u32,
    pub i64const: u32,
    pub i64load: u32,
    pub i64store: u32,
    pub select: u32,
    pub if: u32,
    pub br: u32,
    pub br_if: u32,
    pub br_table: u32,
    pub br_table_per_entry: u32,
    pub call: u32,
    pub call_indirect: u32,
    pub call_per_local: u32,
    pub local_get: u32,
    pub local_set: u32,
    pub local_tee: u32,
    pub global_get: u32,
    pub global_set: u32,
    pub memory_size: u32,
    pub memory_grow: u32,
    pub i64clz: u32,
    pub i64ctz: u32,
    pub i64popcnt: u32,
    pub i64eqz: u32,
    pub i64extendsi32: u32,
    pub i64extendui32: u32,
    pub i32wrapi64: u32,
    pub i64eq: u32,
    pub i64ne: u32,
    pub i64lts: u32,
    pub i64ltu: u32,
    pub i64gts: u32,
    pub i64gtu: u32,
    pub i64les: u32,
    pub i64leu: u32,
    pub i64ges: u32,
    pub i64geu: u32,
    pub i64add: u32,
    pub i64sub: u32,
    pub i64mul: u32,
    pub i64divs: u32,
    pub i64divu: u32,
    pub i64rems: u32,
    pub i64remu: u32,
    pub i64and: u32,
    pub i64or: u32,
    pub i64xor: u32,
    pub i64shl: u32,
    pub i64shrs: u32,
    pub i64shru: u32,
    pub i64rotl: u32,
    pub i64rotr: u32,

}

Describes the weight for all categories of supported wasm instructions.

There there is one field for each wasm instruction that describes the weight to execute one instruction of that name. There are a few exceptions:

If there is a i64 and a i32 variant of an instruction we use the weight of the former for both.

The following instructions are free of charge because they merely structure the wasm module and cannot be spammed without making the module invalid (and rejected): End, Unreachable, Return, Else

The following instructions cannot be benchmarked because they are removed by any real world execution engine as a preprocessing step and therefore don’t yield a meaningful benchmark result. However, in contrast to the instructions mentioned in 2. they can be spammed. We price them with the same weight as the “default” instruction (i64.const): Block, Loop, Nop

We price both i64.const and drop as InstructionWeights.i64const / 2. The reason for that is that we cannot benchmark either of them on its own but we need their individual values to derive (by subtraction) the weight of all other instructions that use them as supporting instructions. Supporting means mainly pushing arguments and dropping return values in order to maintain a valid module.

Fields
version: u32

Version of the instruction weights.

Note

Should be incremented whenever any instruction weight is changed. The reason is that changes to instruction weights require a re-instrumentation in order to apply the changes to an already deployed code. The re-instrumentation is triggered by comparing the version of the current schedule with the version the code was instrumented with. Changes usually happen when pallet_contracts is re-benchmarked.

Changes to other parts of the schedule should not increment the version in order to avoid unnecessary re-instrumentations.

fallback: u32

Weight to be used for instructions which don’t have benchmarks assigned.

This weight is used whenever a code is uploaded with [Determinism::Relaxed] and an instruction (usually a float instruction) is encountered. This weight is not used if a contract is uploaded with [Determinism::Enforced]. If this field is set to 0 (the default) only deterministic codes are allowed to be uploaded.

i64const: u32i64load: u32i64store: u32select: u32if: u32br: u32br_if: u32br_table: u32br_table_per_entry: u32call: u32call_indirect: u32call_per_local: u32local_get: u32local_set: u32local_tee: u32global_get: u32global_set: u32memory_size: u32memory_grow: u32i64clz: u32i64ctz: u32i64popcnt: u32i64eqz: u32i64extendsi32: u32i64extendui32: u32i32wrapi64: u32i64eq: u32i64ne: u32i64lts: u32i64ltu: u32i64gts: u32i64gtu: u32i64les: u32i64leu: u32i64ges: u32i64geu: u32i64add: u32i64sub: u32i64mul: u32i64divs: u32i64divu: u32i64rems: u32i64remu: u32i64and: u32i64or: u32i64xor: u32i64shl: u32i64shrs: u32i64shru: u32i64rotl: u32i64rotr: u32