Struct scrypto_test::prelude::ManifestBuilder

pub struct ManifestBuilder {
    registrar: ManifestNameRegistrar,
    instructions: Vec<InstructionV1>,
    blobs: BTreeMap<Hash, Vec<u8>>,
}

A manifest builder for use in tests.

Note - if you break invariants of the manifest builder (eg resolve a bucket before it’s been created, or pass an invalid parameter), this builder will panic. As such, it’s only designed for use in test code, or where the inputs are trusted.

Simple use case:

let manifest = ManifestBuilder::new()
    .lock_fee_from_faucet()
    .withdraw_from_account(from_account_address, XRD, dec!(1))
    .take_from_worktop(XRD, dec!(1), "xrd")
    .try_deposit_or_abort(to_account_address, None, "xrd")
    .build();

Intermediate use case, where we need to pass a bucket into a component:

let manifest = ManifestBuilder::new()
    .lock_fee_from_faucet()
    .withdraw_from_account(from_account_address, XRD, dec!(1))
    .take_from_worktop(XRD, dec!(1), "xrd")
    .call_function_with_name_lookup(
        package_address,
        "SomeBlueprint",
        "some_function",
        |lookup| (
            lookup.bucket("xrd"),
        ),
    )
    .build();

Advanced use case, where we need to generate a collision-free bucket name:

let mut builder = ManifestBuilder::new()
    .lock_fee_from_faucet()
    .get_free_xrd_from_faucet();
for _ in 0..32 {
    // The generate_bucket_name method generates a new bucket name starting with
    // "transfer" that doesn't collide with any previously used bucket names
    let bucket_name = builder.generate_bucket_name("transfer");
    builder = builder
        .take_from_worktop(XRD, "0.001", &bucket_name)
        .try_deposit_or_abort(to_account_address, None, bucket_name);
}
let manifest = builder.build();

Fields
registrar: ManifestNameRegistrarinstructions: Vec<InstructionV1>blobs: BTreeMap<Hash, Vec<u8>>
Implementations

impl ManifestBuilder

pub fn new() -> ManifestBuilder

Starts a new transaction builder.

pub fn name_lookup(&self) -> ManifestNameLookup

pub fn then( self, next: impl FnOnce(ManifestBuilder) -> ManifestBuilder ) -> ManifestBuilder

pub fn with_name_lookup( self, next: impl FnOnce(ManifestBuilder, ManifestNameLookup) -> ManifestBuilder ) -> ManifestBuilder

pub fn with_bucket( self, bucket: impl ExistingManifestBucket, next: impl FnOnce(ManifestBuilder, ManifestBucket) -> ManifestBuilder ) -> ManifestBuilder

pub fn bucket(&self, name: impl AsRef<str>) -> ManifestBucket

pub fn proof(&self, name: impl AsRef<str>) -> ManifestProof

pub fn named_address(&self, name: impl AsRef<str>) -> ManifestAddress

pub fn address_reservation( &self, name: impl AsRef<str> ) -> ManifestAddressReservation

pub fn generate_bucket_name(&self, prefix: impl Into<String>) -> String

Generates an unused bucket name with the given prefix. This should be used when you are programmatically generating buckets, and need to generate bucket names which do not clash.

pub fn generate_proof_name(&self, prefix: impl Into<String>) -> String

Generates an unused proof name with the given prefix. This should be used when you are programmatically generating proofs, and need to generate names which do not clash.

pub fn generate_address_reservation_name( &self, prefix: impl Into<String> ) -> String

Generates an unused address reservation name with the given prefix. This should be used when you are programmatically generating address reservations, and need to generate names which do not clash.

pub fn generate_address_name(&self, prefix: impl Into<String>) -> String

Generates an unused address name with the given prefix. This should be used when you are programmatically generating named addresses, and need to generate names which do not clash.

pub fn object_names(&self) -> ManifestObjectNames

pub fn add_blob(&mut self, blob: Vec<u8>) -> ManifestBlobRef

Example usage:

let manifest = ManifestBuilder::new()
    .withdraw_from_account(from_account_address, XRD, dec!(1))
    // ...
    .then(|mut builder| {
        let code_blob_ref = builder.add_blob(vec![]);
        builder
            .call_function(
                package_address,
                "my_blueprint",
                "func_name",
                manifest_args!(code_blob_ref),
            )
    })
    .build();

pub fn add_instruction_advanced( self, instruction: InstructionV1 ) -> (ManifestBuilder, NewSymbols)

Only for use in advanced use cases. Returns all the created symbols as part of the instruction

pub fn take_all_from_worktop( self, resource_address: impl ResolvableResourceAddress, new_bucket: impl NewManifestBucket ) -> ManifestBuilder

Takes resource from worktop.

pub fn take_from_worktop( self, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal, new_bucket: impl NewManifestBucket ) -> ManifestBuilder

Takes resource from worktop, by amount.

pub fn take_non_fungibles_from_worktop( self, resource_address: impl ResolvableResourceAddress, ids: impl IntoIterator<Item = NonFungibleLocalId>, new_bucket: impl NewManifestBucket ) -> ManifestBuilder

Takes resource from worktop, by non-fungible ids.

pub fn return_to_worktop( self, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

Adds a bucket of resource to worktop.

pub fn assert_worktop_contains( self, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

Asserts that worktop contains resource.

pub fn assert_worktop_contains_any( self, resource_address: impl ResolvableResourceAddress ) -> ManifestBuilder

Asserts that worktop contains resource.

pub fn assert_worktop_contains_non_fungibles( self, resource_address: impl ResolvableResourceAddress, ids: impl IntoIterator<Item = NonFungibleLocalId> ) -> ManifestBuilder

Asserts that worktop contains resource.

pub fn pop_from_auth_zone( self, new_proof: impl NewManifestProof ) -> ManifestBuilder

Pops the most recent proof from auth zone.

pub fn push_to_auth_zone( self, proof: impl ExistingManifestProof ) -> ManifestBuilder

Pushes a proof onto the auth zone

pub fn create_proof_from_auth_zone_of_amount( self, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal, new_proof: impl NewManifestProof ) -> ManifestBuilder

Creates proof from the auth zone by amount.

pub fn create_proof_from_auth_zone_of_non_fungibles( self, resource_address: impl ResolvableResourceAddress, ids: impl IntoIterator<Item = NonFungibleLocalId>, new_proof: impl NewManifestProof ) -> ManifestBuilder

Creates proof from the auth zone by non-fungible ids.

pub fn create_proof_from_auth_zone_of_all( self, resource_address: impl ResolvableResourceAddress, new_proof: impl NewManifestProof ) -> ManifestBuilder

Creates proof from the auth zone

pub fn create_proof_from_bucket_of_amount( self, bucket: impl ExistingManifestBucket, amount: impl ResolvableDecimal, new_proof: impl NewManifestProof ) -> ManifestBuilder

Creates proof from a bucket. The bucket is not consumed by this process.

pub fn create_proof_from_bucket_of_non_fungibles( self, bucket: impl ExistingManifestBucket, ids: impl IntoIterator<Item = NonFungibleLocalId>, new_proof: impl NewManifestProof ) -> ManifestBuilder

Creates proof from a bucket. The bucket is not consumed by this process.

pub fn create_proof_from_bucket_of_all( self, bucket: impl ExistingManifestBucket, new_proof: impl NewManifestProof ) -> ManifestBuilder

Creates proof from a bucket. The bucket is not consumed by this process.

pub fn clone_proof( self, proof: impl ExistingManifestProof, new_proof: impl NewManifestProof ) -> ManifestBuilder

Clones a proof.

pub fn allocate_global_address( self, package_address: impl ResolvablePackageAddress, blueprint_name: impl Into<String>, new_address_reservation_name: impl Into<String>, new_address_name: impl Into<String> ) -> ManifestBuilder

pub fn drop_proof(self, proof: impl ExistingManifestProof) -> ManifestBuilder

Drops a proof.

pub fn drop_all_proofs(self) -> ManifestBuilder

Drops all proofs.

pub fn drop_named_proofs(self) -> ManifestBuilder

Drops named proofs.

pub fn drop_auth_zone_signature_proofs(self) -> ManifestBuilder

Drops auth zone signature proofs.

pub fn drop_auth_zone_regular_proofs(self) -> ManifestBuilder

Drops auth zone regular proofs.

pub fn drop_auth_zone_proofs(self) -> ManifestBuilder

Drop auth zone proofs.

pub fn create_fungible_resource( self, owner_role: OwnerRole, track_total_supply: bool, divisibility: u8, resource_roles: FungibleResourceRoles, metadata: ModuleConfig<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, initial_supply: Option<Decimal> ) -> ManifestBuilder

Creates a fungible resource

pub fn create_non_fungible_resource<T, V>( self, owner_role: OwnerRole, id_type: NonFungibleIdType, track_total_supply: bool, resource_roles: NonFungibleResourceRoles, metadata: ModuleConfig<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, initial_supply: Option<T> ) -> ManifestBuilder
where T: IntoIterator<Item = (NonFungibleLocalId, V)>, V: ManifestEncode + NonFungibleData,

Creates a new non-fungible resource

pub fn create_ruid_non_fungible_resource<T, V>( self, owner_role: OwnerRole, track_total_supply: bool, metadata: ModuleConfig<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, resource_roles: NonFungibleResourceRoles, initial_supply: Option<T> ) -> ManifestBuilder
where T: IntoIterator<Item = V>, V: ManifestEncode + NonFungibleData,

pub fn update_non_fungible_data( self, resource_address: impl ResolvableResourceAddress, id: NonFungibleLocalId, field_name: impl Into<String>, data: impl ManifestEncode ) -> ManifestBuilder

pub fn create_identity_advanced(self, owner_role: OwnerRole) -> ManifestBuilder

pub fn create_identity(self) -> ManifestBuilder

pub fn create_validator( self, key: Secp256k1PublicKey, fee_factor: impl ResolvableDecimal, xrd_payment: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn register_validator( self, validator_address: impl ResolvableComponentAddress ) -> ManifestBuilder

pub fn unregister_validator( self, validator_address: impl ResolvableComponentAddress ) -> ManifestBuilder

pub fn signal_protocol_update_readiness( self, validator_address: impl ResolvableComponentAddress, protocol_version_name: &str ) -> ManifestBuilder

pub fn stake_validator_as_owner( self, validator_address: impl ResolvableComponentAddress, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn stake_validator( self, validator_address: impl ResolvableComponentAddress, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn unstake_validator( self, validator_address: impl ResolvableComponentAddress, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn claim_xrd( self, validator_address: impl ResolvableComponentAddress, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn call_function( self, package_address: impl ResolvablePackageAddress, blueprint_name: impl Into<String>, function_name: impl Into<String>, arguments: impl ResolvableArguments ) -> ManifestBuilder

Calls a scrypto function where the arguments should be one of:

A tuple, such as (), (x,) or (x, y, z)

IMPORTANT: If calling with a single argument, you must include a trailing comma in the tuple declaration. This ensures that the rust compiler knows it’s a singleton tuple, rather than just some brackets around the inner value.

A struct which implements ManifestEncode representing the arguments

manifest_args!(x, y, z)

NOTE: If you need access to named buckets/proofs etc, use call_function_with_name_lookup instead.

pub fn call_function_raw( self, package_address: impl ResolvablePackageAddress, blueprint_name: impl Into<String>, function_name: impl Into<String>, arguments: Value<ManifestCustomValueKind, ManifestCustomValue> ) -> ManifestBuilder

Calls a scrypto function where the arguments are a raw ManifestValue. The caller is required to ensure the ManifestValue is a Tuple.

You should prefer call_function or call_function_with_name_lookup instead.

pub fn call_function_with_name_lookup<T>( self, package_address: impl ResolvablePackageAddress, blueprint_name: impl Into<String>, function_name: impl Into<String>, arguments_creator: impl FnOnce(&ManifestNameLookup) -> T ) -> ManifestBuilder
where T: ResolvableArguments,

Calls a scrypto function where the arguments will be created using the given callback, which takes a lookup (allowing for resolving named buckets, proofs, etc) and returns resolvable arguments.

The resolvable arguments should be one of:

A tuple, such as (), (x,) or (x, y, z)

IMPORTANT: If calling with a single argument, you must include a trailing comma in the tuple declaration. This ensures that the rust compiler knows it’s a singleton tuple, rather than just some brackets around the inner value.

A struct which implements ManifestEncode representing the arguments

manifest_args!(x, y, z)

Example:

let manifest = ManifestBuilder::new()
    .lock_fee_from_faucet()
    .withdraw_from_account(from_account_address, XRD, dec!(1))
    .take_from_worktop(XRD, dec!(1), "xrd_bucket")
    .call_function_with_name_lookup(
        package_address,
        "SomeBlueprint",
        "some_function",
        |lookup| (
            "argument1",
            lookup.bucket("xrd_bucket"),
            dec!("1.3")
        ),
    )
    .build();

pub fn call_method( self, address: impl ResolvableGlobalAddress, method_name: impl Into<String>, arguments: impl ResolvableArguments ) -> ManifestBuilder

Calls a scrypto method where the arguments should be one of:

A tuple, such as (), (x,) or (x, y, z)

IMPORTANT: If calling with a single argument, you must include a trailing comma in the tuple declaration. This ensures that the rust compiler knows it’s a singleton tuple, rather than just some brackets around the inner value.

A struct which implements ManifestEncode representing the arguments

manifest_args!(x, y, z)

NOTE: If you need access to named buckets/proofs etc, use call_method_with_name_lookup instead.

pub fn call_metadata_method( self, address: impl ResolvableGlobalAddress, method_name: impl Into<String>, arguments: impl ResolvableArguments ) -> ManifestBuilder

pub fn call_royalty_method( self, address: impl ResolvableGlobalAddress, method_name: impl Into<String>, arguments: impl ResolvableArguments ) -> ManifestBuilder

pub fn call_direct_access_method( self, address: InternalAddress, method_name: impl Into<String>, arguments: impl ResolvableArguments ) -> ManifestBuilder

pub fn set_owner_role( self, address: impl ResolvableGlobalAddress, rule: impl Into<AccessRule> ) -> ManifestBuilder

pub fn lock_owner_role( self, address: impl ResolvableGlobalAddress ) -> ManifestBuilder

pub fn set_main_role( self, address: impl ResolvableGlobalAddress, role_key: impl Into<RoleKey>, rule: impl Into<AccessRule> ) -> ManifestBuilder

pub fn set_role( self, address: impl ResolvableGlobalAddress, role_module: ModuleId, role_key: impl Into<RoleKey>, rule: impl Into<AccessRule> ) -> ManifestBuilder

pub fn get_role( self, address: impl ResolvableGlobalAddress, role_module: ModuleId, role_key: RoleKey ) -> ManifestBuilder

pub fn call_role_assignment_method( self, address: impl ResolvableGlobalAddress, method_name: impl Into<String>, arguments: impl ResolvableArguments ) -> ManifestBuilder

pub fn call_module_method( self, address: impl ResolvableGlobalAddress, module_id: ModuleId, method_name: impl Into<String>, arguments: impl ResolvableArguments ) -> ManifestBuilder

pub fn call_method_raw( self, address: impl ResolvableGlobalAddress, method_name: impl Into<String>, arguments: Value<ManifestCustomValueKind, ManifestCustomValue> ) -> ManifestBuilder

Calls a scrypto method where the arguments are a raw ManifestValue. The caller is required to ensure the ManifestValue is a Tuple.

You should prefer call_function or call_function_with_name_lookup instead.

pub fn call_method_with_name_lookup<T>( self, address: impl ResolvableGlobalAddress, method_name: impl Into<String>, arguments_creator: impl FnOnce(&ManifestNameLookup) -> T ) -> ManifestBuilder
where T: ResolvableArguments,

Calls a scrypto method where the arguments will be created using the given callback, which takes a lookup (allowing for resolving named buckets, proofs, etc) and returns resolvable arguments.

The resolvable arguments should be one of:

A tuple, such as (), (x,) or (x, y, z)

IMPORTANT: If calling with a single argument, you must include a trailing comma in the tuple declaration. This ensures that the rust compiler knows it’s a singleton tuple, rather than just some brackets around the inner value.

A struct which implements ManifestEncode representing the arguments

manifest_args!(x, y, z)

Example:

let manifest = ManifestBuilder::new()
    .lock_fee_from_faucet()
    .withdraw_from_account(from_account_address, XRD, dec!(1))
    .take_from_worktop(XRD, dec!(1), "xrd_bucket")
    .call_method_with_name_lookup(
        component_address,
        "some_function",
        |lookup| (
            "argument1",
            lookup.bucket("xrd_bucket"),
            dec!("1.3")
        ),
    )
    .build();

pub fn claim_package_royalties( self, package_address: impl ResolvablePackageAddress ) -> ManifestBuilder

pub fn set_component_royalty( self, component_address: impl ResolvableComponentAddress, method: impl Into<String>, amount: RoyaltyAmount ) -> ManifestBuilder

pub fn lock_component_royalty( self, component_address: impl ResolvableComponentAddress, method: impl Into<String> ) -> ManifestBuilder

pub fn claim_component_royalties( self, component_address: impl ResolvableComponentAddress ) -> ManifestBuilder

pub fn set_metadata( self, address: impl ResolvableGlobalAddress, key: impl Into<String>, value: impl ToMetadataEntry ) -> ManifestBuilder

pub fn lock_metadata( self, address: impl ResolvableGlobalAddress, key: impl Into<String> ) -> ManifestBuilder

pub fn freeze_metadata( self, address: impl ResolvableGlobalAddress, key: impl Into<String> ) -> ManifestBuilder

pub fn publish_package_advanced( self, address_reservation: Option<String>, code: Vec<u8>, definition: PackageDefinition, metadata: impl Into<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, owner_role: OwnerRole ) -> ManifestBuilder

Publishes a package.

pub fn publish_package( self, code: Vec<u8>, definition: PackageDefinition ) -> ManifestBuilder

Publishes a package with an owner badge.

pub fn publish_package_with_owner( self, code: Vec<u8>, definition: PackageDefinition, owner_badge: NonFungibleGlobalId ) -> ManifestBuilder

Publishes a package with an owner badge.

pub fn new_token_mutable( self, metadata: ModuleConfig<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, owner_role: AccessRule ) -> ManifestBuilder

Creates a token resource with mutable supply.

pub fn new_token_fixed( self, owner_role: OwnerRole, metadata: ModuleConfig<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, initial_supply: impl ResolvableDecimal ) -> ManifestBuilder

Creates a token resource with fixed supply.

pub fn new_badge_mutable( self, metadata: ModuleConfig<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, owner_role: AccessRule ) -> ManifestBuilder

Creates a badge resource with mutable supply.

pub fn new_badge_fixed( self, owner_role: OwnerRole, metadata: ModuleConfig<KeyValueStoreInit<String, GenericMetadataValue<UncheckedUrl, UncheckedOrigin>>>, initial_supply: impl ResolvableDecimal ) -> ManifestBuilder

Creates a badge resource with fixed supply.

pub fn burn_resource( self, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn burn_from_worktop( self, amount: impl ResolvableDecimal, resource_address: impl ResolvableResourceAddress ) -> ManifestBuilder

pub fn burn_all_from_worktop( self, resource_address: impl ResolvableResourceAddress ) -> ManifestBuilder

pub fn burn_non_fungible_from_worktop( self, non_fungible_global_id: NonFungibleGlobalId ) -> ManifestBuilder

pub fn mint_fungible( self, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

pub fn mint_non_fungible<T, V>( self, resource_address: impl ResolvableResourceAddress, entries: T ) -> ManifestBuilder
where T: IntoIterator<Item = (NonFungibleLocalId, V)>, V: ManifestEncode,

pub fn mint_ruid_non_fungible<T, V>( self, resource_address: impl ResolvableResourceAddress, entries: T ) -> ManifestBuilder
where T: IntoIterator<Item = V>, V: ManifestEncode,

pub fn recall( self, vault_address: InternalAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

pub fn recall_non_fungibles( self, vault_address: InternalAddress, non_fungible_local_ids: impl IntoIterator<Item = NonFungibleLocalId> ) -> ManifestBuilder

pub fn freeze_withdraw(self, vault_id: InternalAddress) -> ManifestBuilder

pub fn unfreeze_withdraw(self, vault_id: InternalAddress) -> ManifestBuilder

pub fn freeze_deposit(self, vault_id: InternalAddress) -> ManifestBuilder

pub fn unfreeze_deposit(self, vault_id: InternalAddress) -> ManifestBuilder

pub fn freeze_burn(self, vault_id: InternalAddress) -> ManifestBuilder

pub fn unfreeze_burn(self, vault_id: InternalAddress) -> ManifestBuilder

pub fn new_account_advanced( self, owner_role: OwnerRole, address_reservation: Option<String> ) -> ManifestBuilder

Creates an account.

pub fn new_account(self) -> ManifestBuilder

pub fn lock_fee_and_withdraw( self, account_address: impl ResolvableComponentAddress, amount_to_lock: impl ResolvableDecimal, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

pub fn lock_fee_and_withdraw_non_fungibles( self, account_address: impl ResolvableComponentAddress, amount_to_lock: impl ResolvableDecimal, resource_address: impl ResolvableResourceAddress, ids: impl IntoIterator<Item = NonFungibleLocalId> ) -> ManifestBuilder

pub fn lock_fee_from_faucet(self) -> ManifestBuilder

Locks a large fee from the faucet.

pub fn lock_standard_test_fee( self, account_address: impl ResolvableComponentAddress ) -> ManifestBuilder

Locks a large fee from the XRD vault of an account.

pub fn lock_fee( self, account_address: impl ResolvableComponentAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

Locks a fee from the XRD vault of an account.

pub fn lock_contingent_fee( self, account_address: impl ResolvableComponentAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

pub fn get_free_xrd_from_faucet(self) -> ManifestBuilder

Locks a large fee from the faucet.

pub fn withdraw_from_account( self, account_address: impl ResolvableComponentAddress, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

Withdraws resource from an account.

pub fn withdraw_non_fungible_from_account( self, account_address: impl ResolvableComponentAddress, non_fungible_global_id: NonFungibleGlobalId ) -> ManifestBuilder

Withdraws a single non-fungible from an account.

pub fn withdraw_non_fungibles_from_account( self, account_address: impl ResolvableComponentAddress, resource_address: impl ResolvableResourceAddress, ids: impl IntoIterator<Item = NonFungibleLocalId> ) -> ManifestBuilder

Withdraws non-fungibles from an account.

pub fn burn_in_account( self, account_address: impl ResolvableComponentAddress, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

Withdraws resource from an account.

pub fn burn_non_fungible_in_account( self, account_address: impl ResolvableComponentAddress, non_fungible_global_id: NonFungibleGlobalId ) -> ManifestBuilder

Burns a single non-fungible from an account.

pub fn burn_non_fungibles_in_account( self, account_address: impl ResolvableComponentAddress, resource_address: impl ResolvableResourceAddress, local_ids: impl IntoIterator<Item = NonFungibleLocalId> ) -> ManifestBuilder

Burns non-fungibles from an account.

pub fn create_proof_from_account_of_amount( self, account_address: impl ResolvableComponentAddress, resource_address: impl ResolvableResourceAddress, amount: impl ResolvableDecimal ) -> ManifestBuilder

Creates resource proof from an account.

pub fn create_proof_from_account_of_non_fungible( self, account_address: impl ResolvableComponentAddress, non_fungible_global_id: NonFungibleGlobalId ) -> ManifestBuilder

Creates resource proof from an account.

pub fn create_proof_from_account_of_non_fungibles( self, account_address: impl ResolvableComponentAddress, resource_address: impl ResolvableResourceAddress, local_ids: impl IntoIterator<Item = NonFungibleLocalId> ) -> ManifestBuilder

Creates resource proof from an account.

pub fn deposit( self, account_address: impl ResolvableComponentAddress, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn deposit_batch( self, account_address: impl ResolvableComponentAddress ) -> ManifestBuilder

pub fn try_deposit_or_abort( self, account_address: impl ResolvableComponentAddress, authorized_depositor_badge: Option<ResourceOrNonFungible>, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn try_deposit_batch_or_abort( self, account_address: impl ResolvableComponentAddress, batch: impl ResolvableBucketBatch, authorized_depositor_badge: Option<ResourceOrNonFungible> ) -> ManifestBuilder

Note - the batch should either be:

ManifestExpression::EntireWorktop,

An array, vec, or btreeset of bucket names or ManifestBuckets, eg ["my_bucket_1", "my_bucket_2"]

An empty, explicitly typed array of strings, eg Vec::<String>::new()

pub fn try_deposit_entire_worktop_or_abort( self, account_address: impl ResolvableComponentAddress, authorized_depositor_badge: Option<ResourceOrNonFungible> ) -> ManifestBuilder

pub fn try_deposit_or_refund( self, account_address: impl ResolvableComponentAddress, authorized_depositor_badge: Option<ResourceOrNonFungible>, bucket: impl ExistingManifestBucket ) -> ManifestBuilder

pub fn try_deposit_batch_or_refund( self, account_address: impl ResolvableComponentAddress, batch: impl ResolvableBucketBatch, authorized_depositor_badge: Option<ResourceOrNonFungible> ) -> ManifestBuilder

Note - the batch should either be:

ManifestExpression::EntireWorktop,

An array, vec, or btreeset of bucket names or ManifestBuckets, eg ["my_bucket_1", "my_bucket_2"]

An empty, explicitly typed array of strings, eg Vec::<String>::new()

pub fn try_deposit_entire_worktop_or_refund( self, account_address: impl ResolvableComponentAddress, authorized_depositor_badge: Option<ResourceOrNonFungible> ) -> ManifestBuilder

pub fn create_access_controller( self, controlled_asset: impl ExistingManifestBucket, primary_role: AccessRule, recovery_role: AccessRule, confirmation_role: AccessRule, timed_recovery_delay_in_minutes: Option<u32> ) -> ManifestBuilder

pub fn build(self) -> TransactionManifestV1

Builds a transaction manifest.

pub fn to_canonical_string( &self, network_definition: &NetworkDefinition ) -> Result<String, DecompileError>