Struct scrypto_test::prelude::SystemService

pub struct SystemService<'a, Y, V>
where
    Y: KernelApi<SystemConfig<V>>,
    V: SystemCallbackObject,
{
    pub api: &'a mut Y,
    pub phantom: PhantomData<V>,
}

Provided to upper layer for invoking lower layer service

Fields
api: &'a mut Yphantom: PhantomData<V>
Implementations

impl<'a, Y, V> SystemService<'a, Y, V>
where Y: KernelApi<SystemConfig<V>>, V: SystemCallbackObject,

pub fn new(api: &'a mut Y) -> SystemService<'a, Y, V>

impl<'a, Y, V> SystemService<'a, Y, V>
where Y: KernelApi<SystemConfig<V>>, V: SystemCallbackObject,

pub fn get_blueprint_default_definition( &mut self, blueprint_id: BlueprintId ) -> Result<Rc<BlueprintDefinition>, RuntimeError>

pub fn load_blueprint_definition( &mut self, package_address: PackageAddress, bp_version_key: &BlueprintVersionKey ) -> Result<Rc<BlueprintDefinition>, RuntimeError>

pub fn prepare_global_address( &mut self, blueprint_id: BlueprintId, global_address: GlobalAddress ) -> Result<GlobalAddressReservation, RuntimeError>

pub fn get_node_type_info( &mut self, node_id: &NodeId ) -> Result<TypeInfoSubstate, RuntimeError>

pub fn get_blueprint_info( &mut self, node_id: &NodeId, module_id: Option<AttachedModuleId> ) -> Result<BlueprintInfo, RuntimeError>

pub fn get_actor_type_target( &mut self ) -> Result<BlueprintTypeTarget, RuntimeError>

pub fn current_actor(&mut self) -> Actor

pub fn get_object_info( &mut self, node_id: &NodeId ) -> Result<ObjectInfo, RuntimeError>

pub fn is_feature_enabled( &mut self, node_id: &NodeId, module_id: Option<AttachedModuleId>, feature: &str ) -> Result<bool, RuntimeError>

impl<'a, Y, V> SystemService<'a, Y, V>
where Y: KernelApi<SystemConfig<V>>, V: SystemCallbackObject,

pub fn validate_bp_generic_args( &mut self, blueprint_interface: &BlueprintInterface, schemas: &IndexMap<SchemaHash, VersionedSchema<ScryptoCustomSchema>>, generic_substitutions: &Vec<GenericSubstitution> ) -> Result<(), TypeCheckError>

Validate that the type substitutions match the generic definition of a given blueprint

pub fn validate_kv_store_generic_args( &mut self, schemas: &IndexMap<SchemaHash, VersionedSchema<ScryptoCustomSchema>>, key: &GenericSubstitution, value: &GenericSubstitution ) -> Result<(), TypeCheckError>

Validate that the type substitutions for a kv store exist in a given schema

pub fn get_payload_schema( &mut self, target: &BlueprintTypeTarget, payload_identifier: &BlueprintPayloadIdentifier ) -> Result<(Rc<VersionedSchema<ScryptoCustomSchema>>, LocalTypeId, bool, bool, SchemaOrigin), RuntimeError>

pub fn validate_blueprint_payload( &mut self, target: &BlueprintTypeTarget, payload_identifier: BlueprintPayloadIdentifier, payload: &[u8] ) -> Result<(), RuntimeError>

Validate that a blueprint payload matches the blueprint’s definition of that payload

pub fn validate_blueprint_kv_collection( &mut self, target: &BlueprintTypeTarget, collection_index: u8, payloads: &[(&Vec<u8>, &Vec<u8>)] ) -> Result<PartitionDescription, RuntimeError>

Validate that a blueprint kv collection payloads match the blueprint’s definition

pub fn validate_kv_store_payload( &mut self, target: &KVStoreTypeTarget, payload_identifier: KeyOrValue, payload: &[u8] ) -> Result<(), RuntimeError>

Validate that a key value payload matches the key value store’s definition of that payload

pub fn get_blueprint_type_schema( &mut self, type_id: &BlueprintTypeIdentifier ) -> Result<(Rc<VersionedSchema<ScryptoCustomSchema>>, ScopedTypeId), RuntimeError>