Trait scrypto_test::prelude::BootStore

pub trait BootStore {
    // Required method
    fn read_substate(
        &self,
        node_id: &NodeId,
        partition_num: PartitionNumber,
        substate_key: &SubstateKey
    ) -> Option<IndexedScryptoValue>;
}

The interface to be used during boot loading This interface is different from the CommitableSubstateStore in that these reads should not be tracked / costed since it will cause a protocol break.

Required Methods

fn read_substate( &self, node_id: &NodeId, partition_num: PartitionNumber, substate_key: &SubstateKey ) -> Option<IndexedScryptoValue>

Read a substate from the store