Trait scrypto_test::prelude::KeyValueKeyPayload

pub trait KeyValueKeyPayload: Sized + AsRef<Self::Content> + AsMut<Self::Content> + From<Self::Content> {
    type Content: KeyValueKeyContentSource<Self>;

    // Required method
    fn into_content(self) -> Self::Content;

    // Provided methods
    fn from_content(inner_content: Self::Content) -> Self { ... }
    fn from_content_source<T>(content: T) -> Self
       where T: KeyValueKeyContentSource<Self> { ... }
}

This trait is intended to be implemented by an explicit new type for for the given { content: T } key of a particular key value collection.

Required Associated Types

type Content: KeyValueKeyContentSource<Self>

Required Methods

fn into_content(self) -> Self::Content

Provided Methods

fn from_content(inner_content: Self::Content) -> Self

fn from_content_source<T>(content: T) -> Self
where T: KeyValueKeyContentSource<Self>,

Object Safety

This trait is not object safe.