Type Alias scrypto_test::prelude::AccessControllerStateV1

pub type AccessControllerStateV1 = AccessControllerSubstate;

Aliased Type

struct AccessControllerStateV1 {
    pub controlled_asset: Vault,
    pub timed_recovery_delay_in_minutes: Option<u32>,
    pub recovery_badge: ResourceAddress,
    pub state: (PrimaryRoleLockingState, PrimaryRoleRecoveryAttemptState, PrimaryRoleBadgeWithdrawAttemptState, RecoveryRoleRecoveryAttemptState, RecoveryRoleBadgeWithdrawAttemptState),
}

Fields
controlled_asset: Vault

A vault where the asset controlled by the access controller lives.

timed_recovery_delay_in_minutes: Option<u32>

The amount of time (in minutes) that it takes for timed recovery to be done. Maximum is 4,294,967,295 minutes which is 8171.5511700913 years. When this is None, then timed recovery can not be performed through this access controller.

recovery_badge: ResourceAddress

The resource address of the recovery badge that will be used by the wallet and optionally by other clients as well.

state: (PrimaryRoleLockingState, PrimaryRoleRecoveryAttemptState, PrimaryRoleBadgeWithdrawAttemptState, RecoveryRoleRecoveryAttemptState, RecoveryRoleBadgeWithdrawAttemptState)

The states of the Access Controller.