Type Alias scrypto_test::prelude::ConsensusManagerState

pub type ConsensusManagerState = ConsensusManagerSubstate;

Aliased Type

struct ConsensusManagerState {
    pub started: bool,
    pub epoch: Epoch,
    pub effective_epoch_start_milli: i64,
    pub actual_epoch_start_milli: i64,
    pub round: Round,
    pub current_leader: Option<u8>,
}

Fields
started: bool

Whether the consensus process has started

epoch: Epoch

The current epoch.

effective_epoch_start_milli: i64

The effective start-time of the epoch. This is used to calculate the effective duration, for the purpose of calculating when to change epoch. This will typically be close to the actual_epoch_start_milli but may differ slightly as it attempts to avoid minor systematic drift in the epoch start time.

actual_epoch_start_milli: i64

The actual start-time of the epoch. This is just saved as a sanity-check for checking divergence between actual and effective.

round: Round

The current round in the epoch.

current_leader: Option<u8>

The current leader - this is used for knowing who was the validator for the following round of transactions