Trait scrypto_test::prelude::FieldContentSource

pub trait FieldContentSource<Payload>: Sizedwhere
    Payload: FieldPayload,{
    // Required method
    fn into_content(self) -> <Payload as FieldPayload>::Content;

    // Provided methods
    fn into_payload(self) -> Payload { ... }
    fn into_locked_substate(self) -> FieldSubstate<Payload> { ... }
    fn into_unlocked_substate(self) -> FieldSubstate<Payload> { ... }
}

This trait is intended to be implemented by types which embody the content of a particular field payload.

Note:

Multiple types might be mappable into the payload, and so implement this trait

This trait is only one way - from value into content

This trait uses a generic, because the same type might be usable as a payload for multiple substates

Required Methods

fn into_content(self) -> <Payload as FieldPayload>::Content

Provided Methods

fn into_payload(self) -> Payload

fn into_locked_substate(self) -> FieldSubstate<Payload>

fn into_unlocked_substate(self) -> FieldSubstate<Payload>