Trait scrypto_test::prelude::KeyValueEntryPayload

pub trait KeyValueEntryPayload: Sized + AsRef<Self::Content> + AsMut<Self::Content> + From<Self::Content> {
    type Content: KeyValueEntryContentSource<Self>;

    // Required method
    fn into_content(self) -> Self::Content;

    // Provided methods
    fn from_content(inner_content: Self::Content) -> Self { ... }
    fn from_content_source<T>(content: T) -> Self
       where T: KeyValueEntryContentSource<Self> { ... }
    fn into_locked_substate(self) -> KeyValueEntrySubstate<Self> { ... }
    fn into_unlocked_substate(self) -> KeyValueEntrySubstate<Self> { ... }
}

This trait is intended to be implemented by an explicit new type for for the given { content: T } payload of a particular key value collection.

Required Associated Types

type Content: KeyValueEntryContentSource<Self>

Required Methods

fn into_content(self) -> Self::Content

Provided Methods

fn from_content(inner_content: Self::Content) -> Self

fn from_content_source<T>(content: T) -> Selfwhere T: KeyValueEntryContentSource<Self>,

fn into_locked_substate(self) -> KeyValueEntrySubstate<Self>

fn into_unlocked_substate(self) -> KeyValueEntrySubstate<Self>