Trait scrypto_test::prelude::KeyValueKeyContentSource

pub trait KeyValueKeyContentSource<Payload>: Sized
where
    Payload: KeyValueKeyPayload,
{
    // Required method
    fn into_content(self) -> <Payload as KeyValueKeyPayload>::Content;

    // Provided method
    fn into_key(self) -> Payload { ... }
}

This trait is intended to be implemented by types which embody the content of a key for a particular key value collection.

Note:

Multiple types might be mappable into the key payload, and so implement this trait

This trait is only one way - from value into content

This trait uses a generic, because the same type might be usable as a key for multiple substates

Required Methods

fn into_content(self) -> <Payload as KeyValueKeyPayload>::Content

Provided Methods

fn into_key(self) -> Payload

Object Safety

This trait is not object safe.