Trait scrypto_test::prelude::ListableSubstateDatabase

pub trait ListableSubstateDatabase {
    // Required method
    fn list_partition_keys(
        &self
    ) -> Box<dyn Iterator<Item = DbPartitionKey> + '_>;
}

A partition listing interface between Track and a database vendor.

Required Methods

fn list_partition_keys(&self) -> Box<dyn Iterator<Item = DbPartitionKey> + '_>

Iterates over all partition keys, in an arbitrary order.