Module scrypto_test::sdk

Re-exports

pub use CreationStrategy::*;

Modules

package 🔒

resource 🔒

Structs

BucketFactory

A factory for buckets that can create them (for testing) through multiple creation strategies

FungibleResourceType

InProgressResourceBuilder

Utility for setting up a new resource, which has building in progress.

NonFungibleResourceType

Package

ProofFactory

A factory for Proofs that can create them (for testing) through multiple creation strategies

ResourceBuilder

Utility for setting up a new resource.

Enums

CreationStrategy

The creation strategy used in the BucketFactory and ProofFactory structs allowing them to either create them by disabling auth and minting or by mocking the bucket or proof.

FactoryResourceSpecifier

Constants

DIVISIBILITY_MAXIMUM

The maximum divisibility supported.

DIVISIBILITY_NONE

Not divisible.

Traits

AnyResourceType

CreateWithNoSupplyBuilder

SetAddressReservationBuilder

SetOwnerBuilder

UpdateAuthBuilder

UpdateMetadataBuilder

PUBLIC TRAITS AND METHODS All public methods first - these all need good rust docs